refactor: Separate map and place list into composable functions
